Add support for RHEL AI 1.4 and provide batching support
Batching in SDG was always disabled. This should have led to 1 HTTP
request per completion request, but it ended up causing all completion
requests to be sent in 1 HTTP request (likely a bug in InstructLab).
Since the recommendation is to use batching anyways, this adds a good
default batch setting.

Relates:
https://issues.redhat.com/browse/RHOAIENG-20555
https://issues.redhat.com/browse/RHOAIENG-20476

Signed-off-by: mprahl <mprahl@users.noreply.github.com>